+++ shared/data/src/main/java/com/ivy/data/repository/mapper/TransactionMapper.kt
@@ -150,12 +150,4 @@ class TransactionMapper @Inject constructor() {
             id = id.value
         )
     }
-
-    fun Transaction.getTransactionType(): TransactionType {
-        return when (this) {
-            is Expense -> TransactionType.EXPENSE
-            is Income -> TransactionType.INCOME
-            is Transfer -> TransactionType.TRANSFER
-        }
-    }
 }
