+++ temp/legacy-code/src/main/java/com/ivy/legacy/domain/action/transaction/TrnsWithDateDivsAct.kt
@@ -33,6 +33,7 @@ class TrnsWithDateDivsAct @Inject constructor(
     )
 }
 
+@Deprecated("Uses legacy Transaction")
 class LegacyTrnsWithDateDivsAct @Inject constructor(
     private val accountDao: AccountDao,
     private val exchangeAct: ExchangeAct
