+++ temp/legacy-code/src/main/java/com/ivy/legacy/domain/pure/transaction/FoldTransactions.kt
@@ -90,6 +90,7 @@ suspend fun <A> sumTrns(
     ).head
 }
 
+@Deprecated("Uses legacy Transaction")
 object LegacyFoldTransactions {
 
     @Pure
